chore: extract the start lock, taking ensure-server.ts off the size allow-list
- What: lock-file handling moves from `src/ensure-server.ts` to
  `src/server/start-lock.ts`. `ensure-server.ts` 258 → **202**, and its entry is
  removed from the file-size allow-list rather than lowered.

- Why: #2833 — the ≤250-line rule. Process locking is a separable concern from
  server lifecycle: the lock answers "is another process already trying to start
  one?", which is true regardless of what starting involves.

- Impact: none. Pure move plus renames (`acquireLock` → `acquireStartLock`), no
  logic changed. Two imports that only the moved code used are dropped.

Verified behaviourally, not just by tsc — the lock acquires, refuses a second
acquire while held, releases, and re-acquires:

  lock path     : /Users/nat/.arra-oracle-v2/oracle-http.lock
  first acquire : true   exists: true
  second acquire: false  (correctly held)
  after release : removed
  re-acquire    : true

My first probe reported `false` for every acquire. That was the probe, not the
code: `start-lock.ts` reads `getDataDir()`, and `configure({ dataDir })` runs at
`ensure-server.ts` module load, which importing the lock module alone never
triggers — so it resolved the default `~/.bun-process-manager`, which does not
exist. A red from your own harness is a hypothesis, not a verdict.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate tests/build/ tests/lifecycle/
tests/process-manager/ tests/smoke/` 61 pass.

Refs #2833
